requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.7
gunicorn==21.2.0
packaging==25.0
Pillow==10.0.1
psycopg2-binary==2.9.7
python-decouple==3.8
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.15.0
tzdata==2023.3
whitenoise==6.6.0
